Fix for input method commit that ends with newline
If the input method event contains a commit text that ends with a
newline, text, the commit string is inserted first. This changes the
current block.

This change makes sure that we apply the formatting changes (including
removing the old preedit text) to the old block in this specific
case.

This is a copy of change 6cece0f43a18dc8c9dbf5bc49ce515714f090725 in
qtbase, since the code is duplicated.

Task-number: QTBUG-67836
Change-Id: I531d3f1dce51d840acc0ef7fda9e7affb3192327
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>