Fix JS ownership of model and delegate properties in QtQuick item views
When assigning a JS owned model or delegate to an item view, we must
ensure that they stay alive as long as the item view. This happens
easily for example when doing something like

    delegate: Qt.createComponent(...)

This patch takes the minimally invasive approach by changing the QObject
parent of such objects.

Task-number: QTBUG-50319
Task-number: QTBUG-51620
Change-Id: Ie6384b8dd93dcdc62d49f64b38173b3fc4ffd3b3
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>