V4: Peephole optimize LoadReg/MoveReg
The following sequence:

    StoreReg rX
    LoadReg rX

Can be optimized by dropping the LoadReg, as the value is still in the
accumulator. Also, the sequence:

    StoreReg rX
    MoveReg rY, rX

Can be optimized to:

    StoreReg rX
    StoreReg rY

This last optimization prevents one load from the JS stack (reading rX).

Both cases are only valid if there is no label on the second
instruction.

Change-Id: Ibd4543459e1eab4da55e92248eba544c707c5456
Reviewed-by: Lars Knoll <lars.knoll@qt.io>