Clean up QQmlOpenMetaObject property data structure
Use a vector instead of a list for a more compact storage - random
access is a more frequent operation than appending/removal. Also use
a struct instead of a QPair for increased readability of the code and
encapsulate read and write operations.

Clean up the internal API to avoid unnecessary mutable property
extraction. This also helps to ensure that in all cases but intended
mutation we end up using the property getter that returns a QVariant by
value, something that will allow returning a null variant when QObject
tracking gets fixed in the next patch.

Change-Id: I563a930fe959b7636f9e9dc88a28cdcefc196707
Reviewed-by: Lars Knoll <lars.knoll@qt.io>