BorderImage: fix binding loop warning when binding to border properties
Give each border property its own change signal, but keep using the
borderChanged signal in the way it was being used previously.

It's not possible to auto test this, as there's no way to check
for binding loop warnings and the use case that fails doesn't cause
a noticeable difference in any properties that we could test.

Task-number: QTBUG-67901
Change-Id: Id5f029bc738ed15477c8d99ab8665ce78afbe5cf
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>