Exit test executable with failure if specified test functions are not found
If QtQuickTest test executable was given filters that didn't match in any of the
qml files run, it exited successfully.

In combination with a Coin bug when the test function name contains a space,
this allowed failing tests to pass the CI. If the test function fails, the
repeat attempts would pass two arguments to the executable. Neither of those
hit, but that wasn't considered a problem.

Check that all of the test functions named on the command line are actually
executed during the whole run and otherwise exit with a non-zero exit code.

I assume there's no duplicates in the whole names of test functions scoped with
testcase names.

Task-number: QTBUG-68197
Change-Id: Icf7fe263945403f02920522dfd187aeb76b7cb3c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Kalle Viironen <kalle.viironen@qt.io>
Reviewed-by: Ville Voutilainen <ville.voutilainen@qt.io>