Add test case for delete inside a for-in loop
We had a bug where deleting a property on an object that is being
iterated on via a for-in loop would end up iterating too few times
because the iterator compared its progress against the size of the
internal class, which shrunk with each delete call. Commit
ea164ca4a8ec1e5c568ab82c0c4256a841f77bf0 fixed that by retaining the IC
size but just making the deleted members inaccessible, thereby
"accidentally" fixing QTBUG-40021. This change adds a test for the bug
to ensure that we don't regress.

Task-number: QTBUG-40021
Change-Id: Ie75906b665036c311256d86289c158ee0d29f8b7
Reviewed-by: Lars Knoll <lars.knoll@qt.io>