TableView: remove using a floatingPointMargin
The floatingPointMargin was added at an early stage to avoid
that we ended up in a locked situation where we loaded
and unloaded the same edge continuously without being able to
exit the loading loop. This was done wrongly because we didn't
take spacing into account in QQuickTableViewPrivate::canLoadTableEdge
(spacing was usually one pixel, coincidentally the same as the
floatingPointMargin). This has now been fixed in a previous patch, so
we can remove the faulty floatingPointMargin as well.

Change-Id: I3caa7808298a1954a3c02f609bad98c3c2c2426a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>