Fix external styles in static builds
This fixes a regression caused by d2897a6c. For example, running a
statically built Gallery example with the Material style resulted to:

    QQmlApplicationEngine failed to load component
    qrc:/gallery.qml:58 Type ApplicationWindow unavailable
    file:///path/to/QtQuick/Controls.2/Material/qmldir:-1 module "QtQuick.Controls.Material" plugin "qtquickcontrols2materialstyleplugin" not found

Even if we install .qml files in static builds (so that people can copy
them as a starting point from a Qt for iOS installation), those files
must not be loaded in static builds. Previously this case was handled
by the QT_STATIC guard in the former typeUrl(), but the special case
was missed while taking QQuickStyleSelector into use for the internal
types in d2897a6c.

Change-Id: Ie4d7956c7eccb0f4e67b6f3a2b5368437636fa78
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>