Fix QQmlExtensionPlugin::baseUrl() in static builds
Don't assume the file system. Everything is bundled into resources
in static builds. For example, for the QML modules in Qt, the base URL
was "file::/qt-project.org/imports/QtFoo/Bar". This patch fixes it so
that the base URL becomes "qrc:/qt-project.org/imports/QtFoo/Bar".

This fix removes the need for this type of special casing in QQC2 when
resolving the URL for qmlRegisterType() for internal QML composite
types:

    #ifdef QT_STATIC
        QString url = QLatin1String("qrc") + baseUrl().path();
    #else
        QString url = baseUrl().toString();
    #endif

The piece of code originates from QQC2 commit(s) e3c52b2 (and 199962a)
that were presumably fixing support for static builds.

Change-Id: I4fff354c3ed5309843e9122a0175593f4fbb5661
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>