Fix race condition in async image response handling
It may happen that the finished() signal is emitted from a super quick
image loading thread in the implementation before we had a chance to
connect to the signal in Qt Quick. In that case, record the emission
through a private connection established in the constructor.

The unit test simulates this scenario by emitting the signal before
returning the object, but in a real world scenario it may happen
that the response object is used in a separate thread that ends up
emitting the signal. That may also happen before the thread that will
try to connect to the signal runs.

Task-number: QTBUG-59601
Change-Id: I14de9719db47f4bca2e8a2c6659343cfb4deb526
Reviewed-by: Albert Astals Cid <albert.astals.cid@kdab.com>