qv4arrayobject: Implement Array.of from ES7
One small difficulty: We can't accurately tell whether a thisObject is a
constructor or not, so right now we're just swallowing all exceptions
and creating an array if they occur.

This isn't correct, but it isn't trivial to fix, either: I think we
would have to somehow mark our builtins that are constructors, and allow
those, but use the array fallback for other builtins, calling user functions
and bubbling up if they throw.

This is probably good enough for the time being though, as writing
something like:

    function Test() { throw "Foo"; }

    Array.of.call(Test)

... is probably not very likely, compared to more "usual" use.

Change-Id: Ied341a7fa9c3a7fd907e2815c019bc431171ce62
Reviewed-by: Lars Knoll <lars.knoll@qt.io>