Fix items in layouts not being rendered when layers are used
A layer effect is apparently considered a child item of the layout
that manages the item, and this causes issues when layouting.

As was done for positioners in 865b4ec8, exclude
any items for which QQuickItemPrivate::isTransparentForPositioner()
returns true from layouting. As that commit made QQuickItemLayer set it
to true when the layer is enabled, QQuickShaderEffectSource will be
excluded from layouting.

Task-number: QTBUG-63269
Change-Id: I463ff8d0cea0dfd6c4273f376de347971040d3d0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>