Fix creation of object literals
Our method to create object literals wasn't compliant with the
ES7 spec, as we would in some cases re-order the properties.

This violated the spec which required properties to be created
in order, so that for-of would also iterate over them in creation
order.

As a nice side effect, this simplifies the code and gets a couple
of test cases using computed property names to pass.

Task-number: QTBUG-62512
Change-Id: I6dfe004357c5d46a0890027f4fd9e2d1e1a2a17a
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>