Fix truncating of doubles to integers
When analyzing the data flow and determining that a double is or'ed with
0, we correctly assume a cast to an integer and insert a convert
instruction in the IR instead of the actual bitwise or.. The JIT
implements that, but the interpreter doesn't, resulting in incorrect
values.

This problem is specific to the SSA optimizer and the underlying isel
requirements in 5.9 and does not apply to the design of the code
generator in 5.11 and onwards.

Task-number: QTBUG-68640
Change-Id: I954086b118b022bc6a8e00f17517c2d378825613
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
Reviewed-by: Erik Verbruggen <erik.verbruggen@qt.io>