QQmlAdaptorModel: prepare QQmlListProperty for delegate recycling
This patch will enable delegate item recycling for QQmlListProperty.
So if you e.g set "model: someItem.children" on a TableView, the delegates
can be recycled.

The patch will override the notify function. This function is called from
QQmlDelegateModel whenever the modelData that the delegate item
depends on, changes. This again is needed to trigger all the modelData
bindings in the item to update.

Note that this function will only be used for recycling items. The model
classes don't listen for changes done to a QQmlListProperty, meaning that
the function will otherwise never be called (which is why it was never
implemented, I guess).

Auto testing this will be included with the patches that enables
delegate recycling for TableView.

Change-Id: I34161a998a4581798b692da69164ef29032d6e4f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>