QQuickControl: only set hasBackgroundWidth/Height when necessary
The following binding in the Fusion style's TabButton.qml
implementation causes the last tab's background not to cover its full
width:

    height: control.height - (control.checked ? 0 : 2)

It seems that this line is executed upon the completion of
deferred execution of the background, which somehow results in the
early return check in itemGeometryChanged() not being executed.
After that, there are these two lines:

    extra.value().hasBackgroundWidth = p->widthValid;
    extra.value().hasBackgroundHeight = p->heightValid;

Even though this was just a change in height, hasBackgroundWidth was
set to true early on (when the width was 8; the combined left and right
padding), preventing the control from setting a width on the background
in the future.

This patch fixes the issue by only setting hasBackgroundWidth if the
width was changed, and only setting hasBackgroundHeight if the height
was changed.

Task-number: QTBUG-68556
Change-Id: I4c7dbc60d8e73c60c025e5d6c65f3917e6e4ea08
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>