Attempt to stabilize tst_qquickwindow::openglContextCreatedSignal
The test creates a QQuickWindow, calls show() on it and then waits with
qWaitForWindowExposed() to return true. The expectation of the test is
that when that call returns, an expose event was also successfully
delivered to the window and the scene graph's render loop has attempted
rendering the window contents - using *RenderLoop*::renderWindow(),
which would - in the case of the gui thread loop - create the GL
context. This expectation is expressed using a QVERIFY on a signal spy
that observes the emission of the GL context created signal of
QQuickWindow.

There is no guarantee that by the time qWaitForWindowExposed returned
true, an expose event was delivered to the window. Therefore let's use
QTRY_VERIFY.

Change-Id: I604cbf40dfb7980448f406e26b77678a65faf359
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>