MultiPointHandler::hasCurrentPoints: OK if event has more points
We plan for DragHandler to be able to handle multiple points and thus
to inherit from MultiPointHandler.  But most DragHandlers only handle
single points.  Without this change, the mixer manual test would be
broken: each DragHandler's m_currentPoints will have only one point,
but as soon as the event has multiple points, all the DragHandlers
would return false from wantsPointerEvent, and therefore it becomes
impossible to drag two DragHandlers with two fingers.

This test was just meant as a shortcut/optimization anyway.
But if the event contains fewer points than the handler already knows
that it wants to handle, we can still return early.

Change-Id: Ieea26a09bb9af9ee3c7ff2913f8e90f35c0f070e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>